fix: possible bug on registerAPI
