+++ backend/src/Util/Database/Functions/UsersDBFunctions.hs
@@ -34,7 +34,7 @@ insertAllIntoUsersAppDB :: (ToField a) => [a] -> IO ()
 insertAllIntoUsersAppDB = insertAllIntoTableAppDB "users"
 
 updateAllInUsersAppDB :: [String] -> IO ()
-updateAllInUsersAppDB newValues = updateInUsersAppDB $ zip ["name", "email", "password", "type", "enrollment_number", "university_name"] newValues
+updateAllInUsersAppDB newValues = updateInUsersAppDB $ zip ["name", "email", "password", "type", "enrollment_number", "university_name", "created_at"] newValues
 
 updateAllInUsersWhereAppDB :: [String] -> [(String, String, String)] -> IO ()
 updateAllInUsersWhereAppDB newValues = updateInUsersWhereAppDB $ zip ["name", "email", "password", "type", "enrollment_number", "university_name"] newValues
